Wire NFL weather, CFB player profiles, and CFB power rankings
1. nfl_predictor.py: weather adjustment added to predict(), mirroring
   enhanced_predictor.py's existing NFL/CFB block. Applied before
   save_prediction_factors() so the logged/displayed projected score
   reflects it. CFB intentionally not given this treatment yet --
   weather_model.py's NCAAF_STADIUMS covers only 14 of ~122 FBS teams,
   keyed by full mascot names, not cfb_predictor.py's short team_name
   convention -- a data/naming gap, not a wiring one. Tested live
   against both an outdoor (Buffalo Bills) and dome (Atlanta Falcons)
   matchup; the dome case correctly short-circuits to no adjustment.

2. player_profile.py: added cfb to GAME_LOG_TABLES and STAT_COLS
   (identical schema to nfl, confirmed by direct column comparison --
   straight copy) and to the CLI's --sport choices. Tested live against
   a real cfb_game_log player; game log and notes generate correctly,
   bio degrades to None the same way NFL's already does.

3. ranking_engine.py: added cfb to EFFICIENCY_STAT_MAP, wired to
   cfb_defense_ratings.py's existing get_defense_factor(). Picked
   passing_yards as the single-stat proxy -- its STAT_SQL has no
   combined total-yards/points-allowed stat, and passing yardage is
   more predictive of outcomes than rushing volume, which is skewed by
   game-script. Live-testing this surfaced a real bug: get_defense_factor()
   was returning a Decimal (from Postgres SUM()) instead of the float
   its own signature promised, breaking ranking_engine's arithmetic --
   fixed in cfb_defense_ratings.py by casting before round(). Re-tested
   after the fix: 70 CFB teams ranked with varied, non-neutral
   efficiency scores, confirming real per-team data is flowing through.

+++ cfb_defense_ratings.py
@@ -95,7 +95,13 @@ def get_defense_factors(stat: str, use_cache: bool = True) -> dict:
         if (r["games"] or 0) < MIN_GAMES_FOR_DEFENSE:
             continue
         team_rate = (r["stat_total"] or 0) / r["games"]
-        factors[r["team"]] = round(team_rate / league_rate, 3)
+        # FIXED 2026-09-08: Postgres SUM() returns a Decimal, not a
+        # float, for these columns — found live-testing ranking_engine.py's
+        # new EFFICIENCY_STAT_MAP["cfb"] wiring: `eff_score * 0.20`
+        # downstream raised TypeError mixing Decimal and float. Cast
+        # before round() so this actually returns the float its own
+        # signature already promised, not just for this new caller.
+        factors[r["team"]] = round(float(team_rate / league_rate), 3)
 
     if use_cache:
         _cache[stat] = factors
